+++ nuttx/configs/stm32ldiscovery/src/stm32_buttons.c
@@ -1,7 +1,7 @@
 /****************************************************************************
  * configs/stm32ldiscovery/src/board_buttons.c
  *
- *   Copyright (C) 2013-2014 Gregory Nutt. All rights reserved.
+ *   Copyright (C) 2013-2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
  *
  * Redistribution and use in source and binary forms, with or without
@@ -42,6 +42,7 @@
 #include <stdint.h>
 
 #include <nuttx/arch.h>
+#include <nuttx/board.h>
 #include <arch/board/board.h>
 
 #include "stm32ldiscovery.h"
@@ -49,7 +50,7 @@
 #ifdef CONFIG_ARCH_BUTTONS
 
 /****************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 
 /****************************************************************************
@@ -64,7 +65,7 @@
  * the BUTTON_* definitions in board.h
  */
 
-static const uint16_t g_buttons[NUM_BUTTONS] =
+static const uint32_t g_buttons[NUM_BUTTONS] =
 {
   GPIO_BTN_USER
 };
